requirements.txt
apache-airflow-providers-apache-spark==1.0.3
pymongo==3.12.0